reset dww_patient boilerplate template. Added placeholder screens .tsx files and (main) folder in app. Added Django app named 'api' which will handle api calls from client to interact with database. Renamed dww_backend django project to 'core' to be easier to understand with django app 'api'. Added pycache to gitignore.
