+++ dww_backend/requirements.txt
@@ -1,0 +1,6 @@
+asgiref==3.8.1
+Django==5.1.4
+mysqlclient==2.2.7
+sqlparse==0.5.3
+typing_extensions==4.12.2
+tzdata==2024.2
