+++ dww_patient/package.json
@@ -24,7 +24,7 @@
     "expo-font": "~13.0.2",
     "expo-haptics": "~14.0.0",
     "expo-linking": "~7.0.3",
-    "expo-router": "~4.0.16",
+    "expo-router": "^3.1.0",
     "expo-splash-screen": "~0.29.19",
     "expo-status-bar": "~2.0.0",
     "expo-symbols": "~0.2.0",
